Created sidebar-footer.php
Created php file for footer widget area that I registered in the
functions file earlier.  Now I have created the template for the
widgetized area in the footer. I have added in a condintional statement
which states that this widgeted footer area will only appear if is
activated (that is if we put widgets in it) otherwise it will not
appear. So REMEMBER to make sure in the final wordpress site we add in
widgets to this area to make it appear.