Fixed Broken Theme because of custom-header.php file
Theme was broken because of custom-header.php file. The same code was
repeated twice in this file for some reason. It has been removed now.
But the theme is still broken, which WordPress is indicating is due to
an error in the options.php file.  I am working on this now.