Deleted extra style sheet
I deleted this style sheet before. I don't know why it has been added
again. We don't need it. and it conflicts with other stuff in the other
main theme style sheet.

Even when I deleted this stuff nothing changed on the website.

The sidebar does not fit on my computer screen anymore. and before it
fit perfectly. I am working on figuring this out.